--- a/Project.toml
+++ b/Project.toml
@@ -12,6 +12,7 @@ RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
 [compat]
+ChainRulesCore = "1.15"
 FastGaussQuadrature = "0.4"
 IntervalSets = "0.5, 0.6, 0.7"
 RecipesBase = "1"