--- a/src/_ChainRules.jl
+++ b/src/_ChainRules.jl
@@ -1,1 +1,19 @@
+function ChainRulesCore.frule((_, ΔP, Δi, Δt), ::typeof(bsplinebasis), P::BSplineSpace, i::Integer, t::Real)
+    B = bsplinebasis(P,i,t)
+    ∂B_∂P = @not_implemented "hoge"
+    # ∂B_∂i = NoTangent()
+    ∂B_∂t = bsplinebasis′(P,i,t)
+    return (B, ∂B_∂P*ΔP + ∂B_∂t*Δt)
+end
 
+function ChainRulesCore.rrule(::typeof(bsplinebasis), P::BSplineSpace, i::Integer, t::Real)
+    B = bsplinebasis(P,i,t)
+    # project_t = ProjectTo(t)  # Not sure we need this ProjectTo.
+    function bsplinebasis_pullback(ΔB)
+        P̄ = @not_implemented "fuga"
+        ī = NoTangent()
+        t̄ = bsplinebasis′(P,i,t) * ΔB
+        return (NoTangent(), P̄, ī, t̄)
+    end
+    return (B, bsplinebasis_pullback)
+end