This PR fixes #124.