@hyrodium 2022-09-07
**Before this PR**

```julia
julia> using BasicBSpline, Zygote, BenchmarkTools

julia> P = BSplineSpace{2}(KnotVector(0:12))
BSplineSpace{2, Int64}(KnotVector([0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]))

julia> bsplinebasis(P,1,2.3)
0.24500000000000013

julia> bsplinebasis′(P,1,2.3)
-0.7000000000000002

julia> Zygote.gradient(t -> bsplinebasis(P,1,t), 2.3)
(-0.7000000000000002,)

julia> @benchmark Zygote.gradient(t -> bsplinebasis(P,1,t), 2.3)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  110.038 μs …   6.547 ms  ┊ GC (min … max): 0.00% … 96.05%
 Time  (median):     113.455 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   118.900 μs ± 136.241 μs  ┊ GC (mean ± σ):  2.52% ±  2.16%

    ▇█▃                                                          
  ▂▆███▆▄▄▃▃▃▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂▂▂▁▁▂▂▂▁▂▂ ▃
  110 μs           Histogram: frequency by time          164 μs <

 Memory estimate: 25.12 KiB, allocs estimate: 707.
```

**After this PR**
```julia
julia> using BasicBSpline, Zygote, BenchmarkTools

julia> P = BSplineSpace{2}(KnotVector(0:12))
BSplineSpace{2, Int64}(KnotVector([0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]))

julia> bsplinebasis(P,1,2.3)
0.24500000000000013

julia> bsplinebasis′(P,1,2.3)
-0.7000000000000002

julia> Zygote.gradient(t -> bsplinebasis(P,1,t), 2.3)
(-0.7000000000000002,)

julia> @benchmark Zygote.gradient(t -> bsplinebasis(P,1,t), 2.3)
BenchmarkTools.Trial: 10000 samples with 10 evaluations.
 Range (min … max):  1.815 μs … 649.551 μs  ┊ GC (min … max): 0.00% … 99.33%
 Time  (median):     1.881 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.032 μs ±   9.141 μs  ┊ GC (mean ± σ):  6.34% ±  1.40%

      ▃▆▇█▇▆▄▂▁                                                
  ▁▃▆██████████▇▆▅▅▄▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▃
  1.81 μs         Histogram: frequency by time        2.24 μs <

 Memory estimate: 1.12 KiB, allocs estimate: 26.
```

58.5 times faster :tada: :tada: :tada: :tada: :tada: :tada: